+++ testrunner-typescript/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "testrunner",
+  "name": "testrunner-native",
   "version": "1.0.0",
   "main": "index.js",
   "type": "module",
   "scripts": {
     "build": "rm -rf dist && tsc",
-    "test": "node --experimental-strip-types --test --experimental-test-coverage --experimental-test-module-mocks",
+    "test": "node --test --experimental-test-coverage --experimental-test-module-mocks",
     "test:dev": "npm run build && node --watch --test test/ "
   },
   "directories": {
