+++ testrunner-typescript/test/basics/02-stubs.test.ts
