--- testrunner/src/mock-function/data/verify.ts
@@ -1,7 +1,0 @@
-// input-verified.ts
-import { validateRawInput } from './model';
-
-export function inputVerified(doc: any): boolean {
-  const result = validateRawInput(doc);
-  return result.length === 0 ? true : false;
-}
